requirements.txt
Flask==3.0.0
python-telegram-bot==22.3
python-dotenv==1.0.0
aiohttp==3.9.1